android: Work around broken Bionic getaddrinfo()
The first argument to getaddrifo(), "hostname", is normally allowed
to be NULL.  However, some Bionic implementations erroneously call
strlen() on this parameter without first checking to see if it is
NULL.  Work around this by passing in a hostname of "0.0.0.0".

This problem was seen on the HTC One SV stock ROM, JB 4.1.2.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>